requirements.txt
PyQt5>=5.15.9
pytesseract>=0.3.10
pdf2image>=1.17.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
openpyxl>=3.1.2
PyYAML>=6.0.1
rapidfuzz>=3.5.0

# Build/packaging only (not needed at runtime by end users):
pyinstaller>=6.5.0
